requirements.txt
numpy==1.23.5
scipy==1.9.3
openprompt==1.0.1
pandas==1.5.2
sentence_transformers==2.2.2
torch==1.13.1
python-dotenv
pytorch-lightning==1.9.0
wandb==0.13.7
openai==0.27.8
tiktoken==0.4.0
accelerate==0.28.0
transformers==4.39.0
fschat==0.2.24
line_profiler_pycharm==1.1.0
retry==0.9.2
rank_bm25==0.2.2
fsspec==2023.9.2
peft==0.11.1
